Cock and ball torture (CBT) is a sexual activity involving torture of the male genitals. This may involve directly painful activities, such as wax play, genital spanking, squeezing, ball-busting, genital flogging, urethral play, tickle torture, erotic electrostimulation or even kicking.[1]
The recipient of such activities may receive direct physical pleasure via masochism, or emotional pleasure through knowledge that the play is pleasing to a sadistic dominant.

The practice carries significant health risks.

Testicle cuffs[edit]

A testicle-cuff
A testicle cuff is a ring-shaped device around the scrotum between the body and the testicles such that when closed it does not allow the testicles to pass through it. A common type has two connected cuffs; one for the scrotum and one for the base of the penis. They are one of many different gadgets for restraining the male genitalia. Also a standard padlock may be locked around the scrotum, and without the key the man will not be able to remove it.

Some men enjoy the feeling of being 'owned', and some people enjoy the feeling of 'owning' their partners. Making the man wear testicle cuffs gives a sense that his sexual organs belong to his partner. Equally, there is a certain level of humiliation inherent in the devices, by which some people are sexually aroused. Finally, the cuffs may form part of a sexual fetish of the wearer or his partner.[citation needed]

These are, however, extreme uses of testicle cuffs. More conventionally, the device pulls the testicles down and keeps them in place during stimulation, which has a number of benefits including:

Making the penis appear longer. Pulling the testicles down and away from the base of the penis stretches the skin over the base of the penis and pubic bone, exposing the additional inch or so of shaft that is normally hidden.
Improving sexual arousal. While some men may be aroused by the feeling of being "owned", the physical feeling of the ligaments that suspend the testicles being stretched can have an effect similar to the more common practice of stretching out the legs and pointing the toes.
Preventing the testicles from lifting up so far that they become lodged under the skin immediately adjacent to the base of the penis, a condition which can be very uncomfortable especially if the testicle is then crushed by the slap of skin during thrusting.
Delay or intensify ejaculation by preventing the testicles from rising normally to the "point of no return".